requirements.txt
asgiref==3.5.0
Django==4.0.2
django-tempus-dominus==5.1.2.16
django-widget-tweaks==1.4.12
sqlparse==0.4.2
